requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
transformers==4.51.3
torch==2.7.0
Pillow==10.2.0
python-dotenv==1.0.1
huggingface-hub==0.31.0
requests==2.31.0
pydantic==2.6.1 
torchvision==0.22.0
accelerate==1.6.0
qwen-vl-utils==0.0.11
peft==0.15.2
